backend/requirements.txt
Flask==2.3.2
pymongo==4.4.0
python-dotenv==1.0.0
cerebras-cloud-sdk
spacy==3.6.0
scikit-learn==1.3.0
pandas==2.2.0
gunicorn==20.1.0
pytest==7.4.0
flask-jwt-extended
flask-bcrypt

# Note: After creating venv, run:
# python -m spacy download en_core_web_sm
# Set CEREBRAS_API_KEY in your environment (.env) before using LLM features
